+++ src/api/v2/account_activities.rs
@@ -275,6 +275,15 @@ impl Activity {
     }
   }
 
+  /// The time on which the activity occurred.
+  #[cfg(test)]
+  pub fn time(&self) -> SystemTime {
+    match self {
+      Activity::Trade(trade) => trade.transaction_time,
+      Activity::NonTrade(non_trade) => non_trade.date,
+    }
+  }
+
   /// Convert this activity into a trade activity, if it is of the
   /// corresponding variant.
   pub fn into_trade(self) -> Result<TradeActivity, Self> {
@@ -352,6 +361,14 @@ pub struct ActivityReq {
   /// If `None` all activities will be retrieved.
   #[serde(rename = "activity_types", serialize_with = "optional_vec_to_str")]
   pub types: Option<Vec<ActivityType>>,
+  /// The maximum number of entries to return in the response.
+  ///
+  /// The default and maximum value is 100.
+  #[serde(rename = "page_size")]
+  pub page_size: Option<usize>,
+  /// The ID of the end of your current page of results.
+  #[serde(rename = "page_token")]
+  pub page_token: Option<String>,
 }
 
 
@@ -490,6 +507,7 @@ mod tests {
         ActivityType::Transaction,
         ActivityType::Dividend,
       ]),
+      ..Default::default()
     };
     let activities = client.issue::<Get>(request).await.unwrap();
 
@@ -516,6 +534,7 @@ mod tests {
     let client = Client::new(api_info);
     let request = ActivityReq {
       types: Some(vec![ActivityType::Fill]),
+      ..Default::default()
     };
     let activities = client.issue::<Get>(request).await.unwrap();
 
@@ -542,5 +561,32 @@ mod tests {
     // we parsed something. Note that this may not work for newly
     // created accounts, an order may have to be filled first.
     assert!(!activities.is_empty());
+  }
+
+  /// Check that paging works properly.
+  #[test(tokio::test)]
+  async fn page_activities() {
+    let api_info = ApiInfo::from_env().unwrap();
+    let client = Client::new(api_info);
+    let mut request = ActivityReq {
+      page_size: Some(1),
+      ..Default::default()
+    };
+    let activities = client.issue::<Get>(request.clone()).await.unwrap();
+    // We already make the assumption that there are some activities
+    // available for us to work with in other tests, so we continue down
+    // this road here.
+    assert_eq!(activities.len(), 1);
+    let newest_activity = &activities[0];
+
+    request.page_token = Some(newest_activity.id().to_string());
+
+    let activities = client.issue::<Get>(request.clone()).await.unwrap();
+    assert_eq!(activities.len(), 1);
+    let next_activity = &activities[0];
+
+    // Activities are reported in descending order by time.
+    assert!(newest_activity.time() >= next_activity.time());
+    assert_ne!(newest_activity.id(), next_activity.id());
   }
 }